requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
faiss-cpu>=1.9.0
sentence-transformers>=2.2.0
torch>=2.0.0
spacy>=3.7.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0

